+++ src/app/app.component.ts
@@ -1,13 +1,11 @@
-import { Component, EnvironmentInjector, NgZone, OnInit, effect, inject } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EnvironmentInjector, NgZone, OnInit, effect, inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NotificationService } from './notifications/notification.service';
 import { App, URLOpenListenerEvent } from '@capacitor/app';
 import { ShareInfoType, ShareService } from './share/share.service';
 import { IonApp, IonRouterOutlet } from '@ionic/angular/standalone';
-import { RouterFocusService } from './utils/focus.service';
 import { IntegrityService } from './data/integrity.service';
 import { DbService } from './data/db.service';
-import { ApiService } from './data/api.service';
 import { UiService } from './ui/ui.service';
 import { SettingsService } from './data/settings.service';
 
@@ -16,15 +14,14 @@ import { SettingsService } from './data/settings.service';
   templateUrl: 'app.component.html',
   styleUrls: ['app.component.scss'],
   standalone: true,
+  changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [IonApp, IonRouterOutlet, CommonModule],
 })
 export class AppComponent implements OnInit {
   private notificationService = inject(NotificationService);
   private integrityService = inject(IntegrityService);
-  private api = inject(ApiService);
   private shareService = inject(ShareService);
   private zone = inject(NgZone);
-  private focusService = inject(RouterFocusService);
   private dbService = inject(DbService);
   private ui = inject(UiService);
   private settings = inject(SettingsService);
@@ -65,6 +62,5 @@ export class AppComponent implements OnInit {
 
   stackChanged() {
     this.dbService.getWorkerLogs();
-    this.focusService.focus();
   }
 }