+++ src/app/carousel/carousel.component.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, Component, ElementRef, OnDestroy, output, contentChildren, viewChild } from '@angular/core';
+import { Component, ElementRef, output, contentChildren, viewChild, input, computed, ChangeDetectorRef, inject } from '@angular/core';
 import { CarouselItemComponent } from '../carousel-item/carousel-item.component';
 
 export interface SlideSelect {
@@ -11,20 +11,32 @@ export interface SlideSelect {
   styleUrls: ['./carousel.component.scss'],
   standalone: true,
 })
-export class CarouselComponent implements AfterViewInit, OnDestroy {
+export class CarouselComponent {
   slideChanged = output<SlideSelect>();
   private interval: any;
   private lastValue = -1;
+  private _change = inject(ChangeDetectorRef);
+  enabled = input<boolean>(false);
   children = contentChildren(CarouselItemComponent);
   container = viewChild.required<ElementRef>('container');
 
   public setScrollLeft(left: number) {
     setTimeout(() => {
       this.container().nativeElement.scrollLeft = left;
+      this._change.detectChanges();
     }, 50);
   }
 
-  ngAfterViewInit() {
+  private enableChanged = computed(() => {
+    const enabled = this.enabled();
+    if (enabled) {
+      this.enable();
+    } else {
+      this.disable();
+    }
+  });
+
+  private enable() {
     this.interval = setInterval(() => {
       const padding = parseInt(getComputedStyle(this.container().nativeElement).scrollPaddingLeft);
       const l = this.container().nativeElement.scrollLeft;
@@ -41,7 +53,10 @@ export class CarouselComponent implements AfterViewInit, OnDestroy {
     }, 1000);
   }
 
-  ngOnDestroy() {
-    clearInterval(this.interval);
+  private disable() {
+    if (this.interval) {
+      clearInterval(this.interval);
+      this.interval = undefined;
+    }
   }
 }