+++ src/app/data/db.service.ts
@@ -15,7 +15,7 @@ import {
   DatasetResult,
   Dataset,
   Names,
-  LocationAvailable,
+  LocationHidden,
 } from './models';
 import { call, registerWorker } from './worker-interface';
 import { BurningManTimeZone, clone, data_dust_events, daysUntil, noDate, now, nowAtEvent, static_dust_events } from '../utils/utils';
@@ -50,12 +50,29 @@ export class DbService {
   public restart: WritableSignal<string> = signal('');
   public showPastEvents = false;
   private initialized = false;
-  private locationAvailable: LocationAvailable = { art: false, camps: false, artMessage: '', campMessage: '' };
+  public locationsHidden = signal({ art: false, camps: false, artMessage: '', campMessage: '' });
 
   private prefix = '';
   public overrideDataset: string | undefined;
   private datasetsRead: string[] = [];
   private worker!: Worker;
+
+  public anyLocationsHidden = computed(() => {
+    return this.locationsHidden().art || this.locationsHidden().camps;
+  });
+
+  public allLocationsHidden = computed(() => {
+    return this.locationsHidden().art && this.locationsHidden().camps;
+  });
+
+  public getLocationsHidden(): LocationHidden {
+    return this.locationsHidden();
+  }
+
+  public artLocationsHidden = computed(() => {
+    return this.locationsHidden().art;
+  });
+
 
   public async initWorker(): Promise<void> {
     if (!this.initialized) {
@@ -88,7 +105,7 @@ export class DbService {
 
   public async populate(dataset: string, timezone: string): Promise<DatasetResult> {
     this.initWorker(); // Just to double check
-    const result: DatasetResult = await call(this.worker, DataMethods.Populate, dataset, this.locationAvailable, environment, timezone);
+    const result: DatasetResult = await call(this.worker, DataMethods.Populate, dataset, this.getLocationsHidden(), environment, timezone);
     await this.writeData(dataset, Names.summary, result)
     return result;
   }
@@ -118,16 +135,8 @@ export class DbService {
     }
   }
 
-  public setLocationAvailable(locationAvailable: LocationAvailable) {
-    this.locationAvailable = locationAvailable;
-  }
-
-  public getLocationAvailable(): LocationAvailable {
-    return this.locationAvailable;
-  }
-
-  public locationsHidden(): boolean {
-    return this.locationAvailable.art || this.locationAvailable.camps;
+  public setLocationHidden(locationHidden: LocationHidden) {
+    this.locationsHidden.set(locationHidden);
   }
 
   public async checkEvents(day?: Date): Promise<void> {